+++ reduct_rs/src/bucket.rs
@@ -8,6 +8,9 @@ use crate::http_client::HttpClient;
 use reduct_base::msg::bucket_api::{BucketInfo, BucketSettings, FullBucketInfo};
 use reqwest::Method;
 
+use crate::record::query::QueryBuilder;
+use crate::record::read_record::ReadRecordBuilder;
+use crate::record::WriteRecordBuilder;
 use reduct_base::msg::entry_api::EntryInfo;
 use std::sync::Arc;
 
@@ -29,7 +32,11 @@ impl Bucket {
     ///
     /// Returns an error if the bucket could not be removed.
     pub async fn remove(&self) -> Result<()> {
-        self.http_client.delete(&format!("/b/{}", self.name)).await
+        let request = self
+            .http_client
+            .request(Method::DELETE, &format!("/b/{}", self.name));
+        self.http_client.send_request(request).await?;
+        Ok(())
     }
 
     /// Get the settings of the bucket.
@@ -76,14 +83,47 @@ impl Bucket {
     pub async fn entries(&self) -> Result<Vec<EntryInfo>> {
         Ok(self.full_info().await?.entries)
     }
+
+    /// Create a record to write to the bucket.
+    pub fn write_record(&self, entry: &str) -> WriteRecordBuilder {
+        WriteRecordBuilder::new(
+            self.name.clone(),
+            entry.to_string(),
+            Arc::clone(&self.http_client),
+        )
+    }
+
+    /// Create a record to write to the bucket.
+    pub fn read_record(&self, entry: &str) -> ReadRecordBuilder {
+        ReadRecordBuilder::new(
+            self.name.clone(),
+            entry.to_string(),
+            Arc::clone(&self.http_client),
+        )
+    }
+
+    /// Create a record to write to the bucket.
+    pub fn query(&self, entry: &str) -> QueryBuilder {
+        QueryBuilder::new(
+            self.name.clone(),
+            entry.to_string(),
+            Arc::clone(&self.http_client),
+        )
+    }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use bytes::{Buf, Bytes, BytesMut};
+    use chrono::Duration;
+    use futures_util::{pin_mut, StreamExt};
+    use std::io::Read;
 
     use crate::client::tests::{bucket_settings, client};
     use crate::client::ReductClient;
+
+    use crate::record::Record;
     use reduct_base::error::ErrorCode;
     use rstest::{fixture, rstest};
 
@@ -133,6 +173,184 @@ mod tests {
             bucket.info().await.err().unwrap().status,
             ErrorCode::NotFound
         );
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_bucket_write_record_data(#[future] bucket: Bucket) {
+        let bucket: Bucket = bucket.await;
+        bucket
+            .write_record("test")
+            .timestamp_us(1000)
+            .data(Bytes::from("Hey"))
+            .send()
+            .await
+            .unwrap();
+
+        let record = bucket
+            .read_record("test")
+            .timestamp_us(1000)
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(record.bytes().await.unwrap(), Bytes::from("Hey"));
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_bucket_write_record_stream(#[future] bucket: Bucket) {
+        let chunks: Vec<Result<_>> = vec![Ok("hello"), Ok(" "), Ok("world")];
+
+        let stream = futures_util::stream::iter(chunks);
+
+        let bucket: Bucket = bucket.await;
+        bucket
+            .write_record("test")
+            .timestamp_us(1000)
+            .content_length(11)
+            .stream(Box::pin(stream))
+            .send()
+            .await
+            .unwrap();
+
+        let record = bucket
+            .read_record("test")
+            .timestamp_us(1000)
+            .send()
+            .await
+            .unwrap();
+        assert_eq!(record.bytes().await.unwrap(), Bytes::from("hello world"));
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_read_record(#[future] bucket: Bucket) {
+        let bucket: Bucket = bucket.await;
+        let record = bucket
+            .read_record("entry-1")
+            .timestamp_us(1000)
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(record.timestamp_us(), 1000);
+        assert_eq!(record.content_length(), 12);
+        assert_eq!(record.content_type(), "text/plain");
+        assert_eq!(record.labels().get("bucket"), Some(&"1".to_string()));
+        assert_eq!(record.labels().get("entry"), Some(&"1".to_string()));
+        assert_eq!(record.bytes().await.unwrap(), Bytes::from("Hey entry-1!"));
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_read_record_as_stream(#[future] bucket: Bucket) {
+        let bucket: Bucket = bucket.await;
+        let record = bucket
+            .read_record("entry-1")
+            .timestamp_us(1000)
+            .send()
+            .await
+            .unwrap();
+
+        let mut stream = record.stream_bytes();
+        assert_eq!(
+            stream.next().await.unwrap(),
+            Ok(Bytes::from("Hey entry-1!"))
+        );
+        assert_eq!(stream.next().await, None);
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_head_record(#[future] bucket: Bucket) {
+        let record: Record = bucket
+            .await
+            .read_record("entry-1")
+            .timestamp_us(1000)
+            .head_only(true)
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(record.timestamp_us(), 1000);
+        assert_eq!(record.content_length(), 12);
+        assert_eq!(record.content_type(), "text/plain");
+        assert_eq!(record.labels().get("bucket"), Some(&"1".to_string()));
+        assert_eq!(record.labels().get("entry"), Some(&"1".to_string()));
+    }
+
+    #[rstest]
+    // #[case(true, 10)]
+    // #[case(false, 10)]
+    // #[case(false, 10_000)]
+    #[case(false, 30_000_000)]
+    #[case(false, 100_000_000)]
+    #[tokio::test]
+    async fn test_query(#[future] bucket: Bucket, #[case] head_only: bool, #[case] size: usize) {
+        let bucket: Bucket = bucket.await;
+        let mut bodies: Vec<Vec<u8>> = Vec::new();
+        for i in 0..3usize {
+            let mut content = Vec::new();
+            for mut j in 0..size {
+                content.push(i as u8);
+            }
+            bodies.push(content);
+
+            bucket
+                .write_record("entry-3")
+                .timestamp_us((i as u64) * 1000_000)
+                .data(Bytes::from(bodies[i].clone()))
+                .send()
+                .await
+                .unwrap();
+        }
+
+        let mut query = bucket
+            .query("entry-3")
+            .ttl(Duration::minutes(1))
+            .head_only(head_only)
+            .send()
+            .await
+            .unwrap();
+        pin_mut!(query);
+        let record = query.next().await.unwrap().unwrap();
+        assert_eq!(record.timestamp_us(), 0);
+        assert_eq!(record.content_length(), size);
+        assert_eq!(record.content_type(), "application/octet-stream");
+
+        if !head_only {
+            assert_eq!(
+                record.bytes().await.unwrap(),
+                Bytes::from(bodies[0].clone())
+            );
+        }
+
+        let record = query.next().await.unwrap().unwrap();
+        assert_eq!(record.timestamp_us(), 1000_000);
+        assert_eq!(record.content_length(), size);
+        assert_eq!(record.content_type(), "application/octet-stream");
+
+        if !head_only {
+            assert_eq!(
+                record.bytes().await.unwrap(),
+                Bytes::from(bodies[1].clone())
+            );
+        }
+
+        let record = query.next().await.unwrap().unwrap();
+        assert_eq!(record.timestamp_us(), 2000_000);
+        assert_eq!(record.content_length(), size);
+        assert_eq!(record.content_type(), "application/octet-stream");
+
+        if !head_only {
+            assert_eq!(
+                record.bytes().await.unwrap(),
+                Bytes::from(bodies[2].clone())
+            );
+        }
+
+        assert!(query.next().await.is_none());
     }
 
     #[fixture]