+++ reduct_rs/src/lib.rs
@@ -6,6 +6,10 @@
 mod bucket;
 mod client;
 mod http_client;
+mod record;
 
 pub use bucket::Bucket;
 pub use client::ReductClient;
+pub use record::read_record::ReadRecordBuilder;
+pub use record::write_record::WriteRecordBuilder;
+pub use record::{Labels, Record, RecordStream};